+++ CHANGELOG.md
@@ -86,6 +86,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a way to remix datasets flexibly
 - Added `from_pretrained_transformer_and_instances` constructor to `Vocabulary`
 - `TransformerTextField` now supports `__len__`.
+- Tango steps for keeping only the strings or text fields from a dataset
 
 ### Fixed